feat: refactor WorkItemsViewModel and add WorkItemsUi
Refactor the WorkItemsViewModel to remove unnecessary 
dependencies and introduce the WorkItemsUi composable 
for displaying work items. The changes improve code 
modularity and enhance the UI for better user interaction 
with work items, including search functionality.
