+++ composeApp/src/commonMain/kotlin/net/tactware/nimbus/projects/ui/specific/RepositoriesViewModel.kt
@@ -7,31 +7,41 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import net.tactware.nimbus.gitrepos.bl.GetReposByProjectIdUseCase
 import net.tactware.nimbus.gitrepos.dal.GitRepo
-import net.tactware.nimbus.projects.bl.ProjectUpdater
+import net.tactware.nimbus.projects.dal.entities.ProjectIdentifier
 import org.koin.core.annotation.Factory
 import org.koin.core.annotation.InjectedParam
-import kotlin.uuid.Uuid
 
+/**
+ * ViewModel for displaying repositories for a specific project.
+ */
 @Factory
-class SpecificProjectViewModel(
+class RepositoriesViewModel(
     @InjectedParam
-    projectId: Uuid,
-    projectUpdater: ProjectUpdater,
-    getReposByProjectIdUseCase: GetReposByProjectIdUseCase,
+    private val projectIdentifier: ProjectIdentifier,
+    private val getReposByProjectIdUseCase: GetReposByProjectIdUseCase,
 ) : ViewModel() {
 
     private val _projectGitRepos = MutableStateFlow<List<GitRepo>>(emptyList())
     val projectGitRepos = _projectGitRepos.asStateFlow()
 
+    // Search query
+    private val _searchText = MutableStateFlow("")
+    val searchText = _searchText.asStateFlow()
+
     init {
         viewModelScope.launch {
-            projectUpdater.update(projectId)
-        }
-
-        viewModelScope.launch {
-            getReposByProjectIdUseCase.invoke(projectId).collect {
+            getReposByProjectIdUseCase.invoke(projectIdentifier.id).collect {
                 _projectGitRepos.value = it
             }
         }
+    }
+
+    /**
+     * Updates the search query.
+     * 
+     * @param query The search query
+     */
+    fun updateSearchText(query: String) {
+        _searchText.value = query
     }
 }
