+++ composeApp/src/commonMain/kotlin/net/tactware/nimbus/projects/ui/specific/WorkItemsUi.kt
@@ -1,0 +1,149 @@
+package net.tactware.nimbus.projects.ui.specific
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material3.Divider
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import app.cash.paging.compose.collectAsLazyPagingItems
+import net.tactware.nimbus.projects.dal.entities.ProjectIdentifier
+import net.tactware.nimbus.projects.dal.entities.WorkItem
+import org.koin.compose.viewmodel.koinViewModel
+import org.koin.core.parameter.parametersOf
+
+/**
+ * UI component for displaying work items for a specific project.
+ */
+@Composable
+fun WorkItemsUi(projectIdentifier: ProjectIdentifier) {
+    val viewModel = koinViewModel<WorkItemsViewModel> { parametersOf(projectIdentifier) }
+
+    // Collect states from ViewModel
+    val isSearchMode = viewModel.isSearchMode.collectAsState().value
+    val searchResults = viewModel.searchResults.collectAsState().value
+    val workItemsPaging = viewModel.workItemsPaging.collectAsLazyPagingItems()
+    val isLoading = viewModel.isLoading.collectAsState().value
+
+    Column(modifier = Modifier.padding(8.dp)) {
+        // Search field for work items
+        val searchText = viewModel.searchQuery.collectAsState().value
+        TextField(
+            value = searchText,
+            onValueChange = { 
+                viewModel.updateSearchQuery(it)
+            },
+            label = { Text("Search Work Items") },
+            modifier = Modifier.fillMaxWidth().padding(bottom = 8.dp)
+        )
+
+        // Loading indicator
+        if (isLoading) {
+            Text(
+                "Loading...",
+                style = MaterialTheme.typography.bodyMedium,
+                modifier = Modifier.padding(8.dp)
+            )
+        }
+
+        // Work items header
+        Row(
+            modifier = Modifier.fillMaxWidth().padding(vertical = 8.dp),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Text(
+                "ID",
+                style = MaterialTheme.typography.titleSmall,
+                modifier = Modifier.weight(0.1f)
+            )
+            Text(
+                "Title",
+                style = MaterialTheme.typography.titleSmall,
+                modifier = Modifier.weight(0.5f)
+            )
+            Text(
+                "State",
+                style = MaterialTheme.typography.titleSmall,
+                modifier = Modifier.weight(0.2f)
+            )
+            Text(
+                "Assigned To",
+                style = MaterialTheme.typography.titleSmall,
+                modifier = Modifier.weight(0.2f)
+            )
+        }
+        Divider()
+
+        // Display work items based on mode (search or paging)
+        if (isSearchMode) {
+            // Display search results
+            LazyColumn(
+                modifier = Modifier.weight(1f)
+            ) {
+                items(searchResults) { workItem ->
+                    WorkItemRow(workItem)
+                    Divider()
+                }
+            }
+        } else {
+            // Display paged work items
+            if (workItemsPaging.itemCount == 0 && !isLoading) {
+                // Show loading indicator for initial paging data load
+                Text(
+                    "Loading work items...",
+                    style = MaterialTheme.typography.bodyMedium,
+                    modifier = Modifier.padding(8.dp)
+                )
+            }
+
+            LazyColumn(
+                modifier = Modifier.weight(1f)
+            ) {
+                items(workItemsPaging.itemCount) { index ->
+                    val workItem = workItemsPaging[index]
+                    if (workItem != null) {
+                        WorkItemRow(workItem)
+                        Divider()
+                    }
+                }
+            }
+        }
+    }
+}
+
+/**
+ * Displays a single work item row.
+ */
+@Composable
+private fun WorkItemRow(workItem: WorkItem) {
+    Row(
+        modifier = Modifier.fillMaxWidth().padding(vertical = 8.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Text(
+            workItem.id.toString(),
+            modifier = Modifier.weight(0.1f)
+        )
+        Text(
+            workItem.title,
+            modifier = Modifier.weight(0.5f)
+        )
+        Text(
+            workItem.state,
+            modifier = Modifier.weight(0.2f)
+        )
+        Text(
+            workItem.assignedTo ?: "Unassigned",
+            modifier = Modifier.weight(0.2f)
+        )
+    }
+}
